// web/src/shared/api/generated/model/paymentMetricsResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Elmanhg.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { DailyValueResult } from './dailyValueResult';
import type { Money } from './money';

export interface PaymentMetricsResult {
  from: string;
  to: string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  succeeded: number | string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  failed: number | string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  refunds: number | string;
  revenue: Money;
  refunded: Money;
  netRevenue: Money;
  revenueByDay: DailyValueResult[];
  generatedAt: string;
}