// web/src/shared/api/generated/model/solveRateMetricsResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Elmanhg.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { SolveRateDayResult } from './solveRateDayResult';

export interface SolveRateMetricsResult {
  from: string;
  to: string;
  /** @nullable */
  subjectId: string | null;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  attempts: number | string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  activeStudentDays: number | string;
  /**
     * @nullable
     * @pattern ^-?(?:0|[1-9]\d*)(?:\.\d+)?$
     */
  attemptsPerActiveStudentPerDay: number | string | null;
  daily: SolveRateDayResult[];
  generatedAt: string;
}