// web/src/shared/api/generated/model/studentMetricsResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Elmanhg.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { DailyValueResult } from './dailyValueResult';

export interface StudentMetricsResult {
  from: string;
  to: string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  total: number | string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  newInRange: number | string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  newThisWeek: number | string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  activeToday: number | string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  activeThisMonth: number | string;
  dailyActive: DailyValueResult[];
  generatedAt: string;
}