// web/src/shared/api/generated/model/subscriberMetricsResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Elmanhg.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { Money } from './money';
import type { PlanCountResult } from './planCountResult';

export interface SubscriberMetricsResult {
  from: string;
  to: string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  activeSubscriptions: number | string;
  activeByPlan: PlanCountResult[];
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  churnedInRange: number | string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  churnedThisMonth: number | string;
  monthlyRecurringRevenue: Money;
  generatedAt: string;
}